fix: after register and edit dont update user cache
